Initial commit
So this is a thing I've been working on for a while now. It's a test runner, developed to run exa's integration test suite (and is now actually being used by it, which is why I finally have to make it public).

You define your checks in TOML files, and it runs them, and prints out a report of how well they did. It can check files on the local machine, the results of commands, DNS queries, HTTP queries, users and groups, package databases — pretty much anything I needed. I'm using it for many things myself, so it's time to lift the lid off the box.

The code is in a bit of a state in places. There are missing docs and plenty of unwraps and expects, not to mention Clippy warnings, but the main structure and architecture of the code seems to be firmly set.